Split FrequencyInfo class into its own source file
